undo support
super inefficient because it copies the entire document! yikes.

perhaps we could reuse lines that are the same between successive undos
or have entire sequences instead of line-by-line buffers.

i seem to recall neovim keeping only the sections that are changed,
but it is hard to know in advance what the changes are going to be.

so maybe copy the entire document, but save only the diff lines.